+++ frontend/src/store/app.ts
@@ -1,4 +1,4 @@
-import { useDispatch, useSelector } from 'react-redux'
+import { useDispatch, useSelector, useStore } from 'react-redux'
 import type { store } from './store.js'
 import { createAsyncThunk } from '@reduxjs/toolkit'
 
@@ -12,6 +12,7 @@ export type RootState = ReturnType<typeof store.getState>
 // Use throughout your app instead of plain `useDispatch` and `useSelector`
 export const useAppDispatch = useDispatch.withTypes<AppDispatch>()
 export const useAppSelector = useSelector.withTypes<RootState>()
+export const useAppStore = useStore.withTypes<AppStore>()
 
 export const createAppAsyncThunk = createAsyncThunk.withTypes<{
 	state: RootState
